requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.3
boto3==1.33.1
botocore==1.33.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dnspython==2.5.0
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.13.4
h11==0.14.0
idna==3.6
jmespath==1.0.1
mangum==0.17.0
packaging==24.1
pillow==10.4.0
pipenv==2023.12.1
platformdirs==4.2.0
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
s3transfer==0.8.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
termcolor==2.4.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
virtualenv==20.25.1
